Refactor: move some functions
