+++ examples/adaptive_optimizer/experiment_different_targets.py
@@ -1,18 +1,12 @@
 from datetime import timedelta
 from functools import partial
-from pprint import pprint
-from typing import Callable, Optional, List
+from typing import Optional
 
-import networkx as nx
-from matplotlib import pyplot as plt
-
+from examples.adaptive_optimizer.experiment_setup import run_adaptive_mutations
 from examples.synthetic_graph_evolution.graph_search import graph_search_setup
 from examples.synthetic_graph_evolution.generators import postprocess_nx_graph
 from examples.synthetic_graph_evolution.tree_search import tree_search_setup
-from examples.synthetic_graph_evolution.utils import draw_graphs_subplots, plot_action_values
-from golem.core.adapter.nx_adapter import BaseNetworkxAdapter
 from golem.core.optimisers.genetic.gp_optimizer import EvoGraphOptimizer
-from golem.core.optimisers.genetic.operators.operator import PopulationT
 from golem.core.optimisers.objective import Objective
 from golem.metrics.edit_distance import tree_edit_dist
 from golem.metrics.graph_metrics import *
@@ -35,37 +29,6 @@ def generate_trees(graph_sizes: Sequence[int], node_types: Sequence[str] = ('x',
     trees = [nx.random_tree(n, create_using=nx.DiGraph) for n in graph_sizes]
     trees = [postprocess_nx_graph(g, node_labels=node_types) for g in trees]
     return trees
-
-
-def run_adaptive_mutations(
-        target: nx.DiGraph,
-        objective: Objective,
-        optimizer: EvoGraphOptimizer,
-        visualize: bool = True,
-):
-    stats_action_value_log: List[List[float]] = []
-
-    def log_action_values(next_pop: PopulationT, optimizer: EvoGraphOptimizer):
-        values = optimizer.mutation.agent.get_action_values(obs=None)
-        stats_action_value_log.append(list(values))
-
-    # Setup the logger and run the optimizer
-    optimizer.set_iteration_callback(log_action_values)
-    found_graphs = optimizer.optimise(objective)
-    found_graph = found_graphs[0] if isinstance(found_graphs, Sequence) else found_graphs
-    history = optimizer.history
-    agent = optimizer.mutation.agent
-
-    print('History of action probabilities:')
-    pprint(stats_action_value_log)
-    if visualize:
-        found_nx_graph = BaseNetworkxAdapter().restore(found_graph)
-        draw_graphs_subplots(target, found_nx_graph,
-                             titles=['Target Graph', 'Found Graph'])
-        history.show.fitness_line()
-        plot_action_values(stats_action_value_log, action_tags=agent.actions)
-        plt.show()
-    return stats_action_value_log
 
 
 def run_experiment_graphs(trial_timeout: int = 15, trial_iterations: Optional[int] = 500):
