+++ examples/synthetic_graph_evolution/utils.py
@@ -1,13 +1,11 @@
-from datetime import datetime
-from itertools import chain, product
-from typing import Tuple, Optional, Sequence, Iterable, Any
+from itertools import chain
+from typing import Tuple, Optional, Sequence
 
 import matplotlib.patches as mpatches
 import matplotlib.pyplot as plt
 import networkx as nx
 import numpy as np
 from matplotlib import cm
-from scipy.special import softmax
 
 from golem.core.adapter.nx_adapter import BaseNetworkxAdapter
 from golem.core.optimisers.opt_history_objects.opt_history import OptHistory
@@ -150,24 +148,3 @@ def _get_node_colors_and_labels(graph: nx.Graph,
     handles = [root_legend, default_legend, source_legend]
 
     return colors, labels, handles
-
-
-def plot_action_values(stats: Sequence[Sequence[float]],
-                       action_tags: Optional[Sequence[Any]] = None,
-                       size: float = 5.,
-                       ylim: float = 5.):
-    # Plot stackplot of how action expectations and probabilities changed
-    x = np.arange(len(stats))
-    y = np.array(stats).T
-    y_prob = softmax(y, axis=0)
-
-    labels = [str(action) for action in action_tags]
-    fig, ax = plt.subplots(nrows=1, ncols=2, figsize=(size * 2, size))
-    ax[0].stackplot(x, y, labels=labels)
-    ax[1].stackplot(x, y_prob, labels=labels)
-
-    ax[0].set(ylim=(0, ylim), yticks=np.linspace(0., ylim, 21))
-    ax[1].set(ylim=(0, 1.0), yticks=np.linspace(0., 1., 21))
-
-    if action_tags:
-        ax[0].legend(loc='upper right')
